% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{E_req}
\alias{E_req}
\title{Get the number of rides required to increment to the next Eddington number}
\usage{
E_req(rides)
}
\arguments{
\item{rides}{A vector of mileage, where each element represents a single
ride.}
}
\value{
A named list with the current Eddington number (\code{E}) and the
  number of rides required to increment by one (\code{req}).
}
\description{
Get the number of rides required to increment to the next Eddington number
}
\seealso{
\code{\link{E_num}}, \code{\link{E_sat}}, \code{\link{E_cum}}
}
