% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx.r
\name{get_haversine_distance}
\alias{get_haversine_distance}
\title{Compute the distance between two points using the Haversine formula}
\usage{
get_haversine_distance(
  lat_1,
  long_1,
  lat_2,
  long_2,
  units = c("miles", "kilometers")
)
}
\arguments{
\item{lat_1, long_1, lat_2, long_2}{The coordinates
used to compute the distance.}

\item{units}{The units of the output distance.}
}
\value{
The distance between two points in the requested units.
}
\description{
Uses the Haversine great-circle distance formula to compute the distance
between two latitude/longitude points.
}
\references{
\url{https://en.wikipedia.org/wiki/Haversine_formula}
}
