% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddington-mod.r
\name{EddingtonModule}
\alias{EddingtonModule}
\title{An Rcpp Module for Tracking Eddington Numbers for Cycling}
\description{
An Rcpp Module for Tracking Eddington Numbers for Cycling
}
\examples{
e <- new(Eddington, c(3, 3, 2))
e$current

e$update(c(3, 3, 5))
e$cumulative

}
