% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_haversine_distance}
\alias{get_haversine_distance}
\title{Compute the distance between two points using the Haversine formula}
\usage{
get_haversine_distance(lat_1, long_1, lat_2, long_2, r)
}
\arguments{
\item{lat_1, long_1, lat_2, long_2}{The coordinates
used to compute the distance.}

\item{r}{The radius of the Earth.}
}
\description{
Compute the distance between two points using the Haversine formula
}
\references{
\url{https://en.wikipedia.org/wiki/Haversine_formula}
}
